backend/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
elevenlabs>=0.2.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing dependencies
pytest>=7.0.0